Moves flakey appointment-list e2e tests to separate spec
