+++ src/applications/vaos/tests/e2e/flows/appointment-list.cypress.spec.js
@@ -10,7 +10,6 @@ import {
   mockAppointmentRequestsApi,
   vaosSetup,
   mockFacilitiesApi,
-  mockCancelReasonsApi,
   mockUserTransitionAvailabilities,
   mockClinicApi,
   mockAppointmentApi,
@@ -31,40 +30,6 @@ describe('VAOS appointment list', () => {
       mockLoginApi();
       mockUserTransitionAvailabilities();
       mockVamcEhr();
-    });
-
-    it('community care appointment', () => {
-      const data = [
-        {
-          id: '1',
-          type: 'Appointment',
-          attributes: {
-            id: 1,
-            kind: 'cc',
-            status: 'booked',
-            start: moment().format('YYYY-MM-DDTHH:mm:ss'),
-          },
-        },
-      ];
-
-      mockAppointmentsApi({ data, apiVersion: 2 });
-
-      cy.visit(rootUrl);
-      cy.injectAxe();
-
-      cy.wait(['@v2:get:appointments']);
-
-      cy.get('[data-cy=upcoming-appointment-list-header]').should('exist');
-      cy.get('[data-cy=appointment-list-item]')
-        .contains(/Community Care/i)
-        .first()
-        .click();
-      cy.url().should('include', '/appointments/cc');
-      cy.get('[data-cy=community-care-appointment-details-header]')
-        .should('exist')
-        .contains('Community care');
-
-      cy.axeCheckBestPractice();
     });
 
     it('va appointment', () => {
@@ -314,45 +279,6 @@ describe('VAOS appointment list', () => {
       cy.get('[data-cy=va-video-appointment-details-header]')
         .should('exist')
         .contains('VA Video Connect at home');
-
-      cy.axeCheckBestPractice();
-    });
-
-    it('should allow for canceling of appointments', () => {
-      const data = [
-        {
-          id: '1',
-          type: 'Appointment',
-          attributes: {
-            id: 1,
-            kind: 'clinic',
-            status: 'booked',
-            start: moment().format('YYYY-MM-DDTHH:mm:ss'),
-            cancellable: true,
-          },
-        },
-      ];
-
-      mockAppointmentsApi({ data, apiVersion: 2 });
-      mockFacilityApi({ id: 'vha_442GC', apiVersion: 1 });
-      mockCancelReasonsApi({ facilityId: '983' });
-
-      cy.visit(rootUrl);
-      cy.injectAxe();
-
-      cy.wait(['@v2:get:appointments']);
-
-      cy.get('[data-cy=appointment-list-item]')
-        .contains(/VA Appointment/i)
-        .first()
-        .click();
-      cy.findByText(/Appointment detail/i).should('exist');
-      cy.findByText(/Cancel appointment/i).click();
-      cy.findByText(/Yes, cancel this appointment/i).click();
-      cy.findByTestId('cancel-appointment-SuccessModal').should('exist');
-      cy.contains('button', /Continue/i).click();
-      cy.get('#cancelAppt').should('not.exist');
-      cy.get('.usa-alert-success').should('not.exist');
 
       cy.axeCheckBestPractice();
     });
